Rename dto mapping class
