+++ Database/Search/IMapProductsToDto.cs
@@ -1,0 +1,9 @@
+﻿using Database.Models.DTO;
+
+namespace Database.Search
+{
+    public interface IMapProductsToDto
+    {
+        IList<ProductDto> Map(IQueryable<Models.Product> products);
+    }
+}
