--- Database/Search/IProductsToDtoMapper.cs
@@ -1,9 +1,0 @@
-﻿using Database.Models.DTO;
-
-namespace Database.Search
-{
-    public interface IProductsToDtoMapper
-    {
-        IList<ProductDto> Map(IQueryable<Models.Product> products);
-    }
-}
